Refactor manifest logic into custom hook
Couple things happening together in this commit -- in addition to adding the custom hook, we needed to figure out proper overall structure to accommodate the necessary server and client components.  The app now checks state to ensure the manifest has been loaded before it tries to fetch info from the datastore, so this seems to be meeting our needs for now.